requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.5
livekit==0.11.0
livekit-api==1.0.7
livekit-agents==0.8.0
Pillow==10.4.0
gtts==2.5.0
opencv-python==4.10.0.84
numpy==1.26.4
